# std/math: optimize division with divisors less than a half-limb